+++ tests/Promise/PromiseTest.cpp
@@ -12,9 +12,9 @@ namespace QtPromise
 namespace Tests
 {
 
-const QString ACTION_RESOLVE = "resolve";
-const QString ACTION_REJECT = "reject";
-const QString ACTION_NOTIFY = "notify";
+const QString ACTION_RESOLVE = QStringLiteral("resolve");
+const QString ACTION_REJECT = QStringLiteral("reject");
+const QString ACTION_NOTIFY = QStringLiteral("notify");
 
 /*! Unit tests for the Promise class.
  *
@@ -55,6 +55,11 @@ private Q_SLOTS:
 	void testAllAnySync_data();
 	void testAllAnySync();
 	void testAllAnyInitializerList();
+	void testPromiseDestruction();
+	void testChainDestruction();
+	void testParentDeferredDestruction();
+	void testDelay_data();
+	void testDelay();
 
 private:
 	struct PromiseSpies
@@ -378,7 +383,7 @@ void PromiseTest::testThenVoidCallback()
 
 	QVariantList chainedNotifiedCalls;
 
-	newPromise->then(nullptr, nullptr, [&chainedNotifiedCalls] (const QVariant& progress) {
+	Promise::Ptr chainedNotifyPromise = newPromise->then(nullptr, nullptr, [&chainedNotifiedCalls] (const QVariant& progress) {
 		chainedNotifiedCalls.push_back(progress);
 	});
 
@@ -466,7 +471,7 @@ void PromiseTest::testThenVariantCallback()
 	QVariantList rejectedCalls;
 	QVariantList notifiedCalls;
 
-	newPromise->then([&](const QVariant& value) {
+	Promise::Ptr spyPromise = newPromise->then([&](const QVariant& value) {
 		resolvedCalls.push_back(value);
 	}, [&](const QVariant& reason) {
 		rejectedCalls.push_back(reason);
@@ -510,15 +515,15 @@ void PromiseTest::testThenPromiseCallback_data()
 
 
 	QVariant data("my string value");
-	//                                        // async // action         // callbackAction      // callbackData // expectedOriginalState // expectedChainedState expectedChainedData // expectedResolvedCalls    // expectedRejectedCalls
-	QTest::newRow("sync resolve -> reject")   << false << ACTION_RESOLVE << ACTION_REJECT       << data         << Deferred::Resolved    << Deferred::Rejected   << data             << (QVariantList())         << (QVariantList() << data);
-	QTest::newRow("sync resolve -> resolve")  << false << ACTION_RESOLVE << ACTION_RESOLVE      << data         << Deferred::Resolved    << Deferred::Resolved   << data             << (QVariantList() << data) << (QVariantList());
-	QTest::newRow("sync reject -> resolve")   << false << ACTION_REJECT  << ACTION_RESOLVE      << data         << Deferred::Rejected    << Deferred::Resolved   << data             << (QVariantList() << data) << (QVariantList());
-	QTest::newRow("sync reject -> reject")    << false << ACTION_REJECT  << ACTION_REJECT       << data         << Deferred::Rejected    << Deferred::Rejected   << data             << (QVariantList())         << (QVariantList() << data);
-	QTest::newRow("async resolve -> reject")  << true  << ACTION_RESOLVE << ACTION_REJECT       << data         << Deferred::Resolved    << Deferred::Rejected   << data             << (QVariantList())         << (QVariantList() << data);
-	QTest::newRow("async resolve -> resolve") << true  << ACTION_RESOLVE << ACTION_RESOLVE      << data         << Deferred::Resolved    << Deferred::Resolved   << data             << (QVariantList() << data) << (QVariantList());
-	QTest::newRow("async reject -> resolve")  << true  << ACTION_REJECT  << ACTION_RESOLVE      << data         << Deferred::Rejected    << Deferred::Resolved   << data             << (QVariantList() << data) << (QVariantList());
-	QTest::newRow("async reject -> reject")   << true  << ACTION_REJECT  << ACTION_REJECT       << data         << Deferred::Rejected    << Deferred::Rejected   << data             << (QVariantList())         << (QVariantList() << data);
+	//                                        // async // action         // callbackAction      // callbackData // expectedOriginalState // expectedChainedState // expectedChainedData // expectedResolvedCalls    // expectedRejectedCalls
+	QTest::newRow("sync resolve -> reject")   << false << ACTION_RESOLVE << ACTION_REJECT       << data         << Deferred::Resolved    << Deferred::Rejected   << data                << (QVariantList())         << (QVariantList() << data);
+	QTest::newRow("sync resolve -> resolve")  << false << ACTION_RESOLVE << ACTION_RESOLVE      << data         << Deferred::Resolved    << Deferred::Resolved   << data                << (QVariantList() << data) << (QVariantList());
+	QTest::newRow("sync reject -> resolve")   << false << ACTION_REJECT  << ACTION_RESOLVE      << data         << Deferred::Rejected    << Deferred::Resolved   << data                << (QVariantList() << data) << (QVariantList());
+	QTest::newRow("sync reject -> reject")    << false << ACTION_REJECT  << ACTION_REJECT       << data         << Deferred::Rejected    << Deferred::Rejected   << data                << (QVariantList())         << (QVariantList() << data);
+	QTest::newRow("async resolve -> reject")  << true  << ACTION_RESOLVE << ACTION_REJECT       << data         << Deferred::Resolved    << Deferred::Rejected   << data                << (QVariantList())         << (QVariantList() << data);
+	QTest::newRow("async resolve -> resolve") << true  << ACTION_RESOLVE << ACTION_RESOLVE      << data         << Deferred::Resolved    << Deferred::Resolved   << data                << (QVariantList() << data) << (QVariantList());
+	QTest::newRow("async reject -> resolve")  << true  << ACTION_REJECT  << ACTION_RESOLVE      << data         << Deferred::Rejected    << Deferred::Resolved   << data                << (QVariantList() << data) << (QVariantList());
+	QTest::newRow("async reject -> reject")   << true  << ACTION_REJECT  << ACTION_REJECT       << data         << Deferred::Rejected    << Deferred::Rejected   << data                << (QVariantList())         << (QVariantList() << data);
 }
 
 /*! \test Tests the Promise::then() method with a callback returning Promise::Ptr.
@@ -556,7 +561,7 @@ void PromiseTest::testThenPromiseCallback()
 	QVariantList resolvedCalls;
 	QVariantList rejectedCalls;
 
-	newPromise->then([&](const QVariant& value) {
+	Promise::Ptr spyPromise = newPromise->then([&](const QVariant& value) {
 		resolvedCalls.push_back(value);
 	}, [&](const QVariant& reason) {
 		rejectedCalls.push_back(reason);
@@ -651,7 +656,8 @@ void PromiseTest::testThenNotify()
 	originalDeferred->notify(QVariant("seventh notify"));
 	notifyDeferred->notify(QVariant("eighth notify"));
 
-	QTRY_VERIFY(spies.notified.empty());
+	QTest::qWait(100);
+	QVERIFY(spies.notified.empty());
 
 	// Avoid warning
 	notifyDeferred->resolve();
@@ -672,9 +678,9 @@ void PromiseTest::testThenNotifyPromiseCallback_data()
 	QTest::newRow("resolve") << (QList<bool>{} << true)
 	                         << (QVariantList{} << notifyData)
 	                         << (QVariantList{} << notifyData);
-	QTest::newRow("reject") << (QList<bool>{} << false)
-	                        << (QVariantList{} << notifyData)
-	                        << QVariantList();
+	QTest::newRow("reject")  << (QList<bool>{} << false)
+	                         << (QVariantList{} << notifyData)
+	                         << QVariantList();
 	QTest::newRow("resolve, reject") << (QList<bool>{} << true << false)
 	                                 << (QVariantList{} << notifyData << notifyData2)
 	                                 << (QVariantList{} << notifyData);
@@ -1042,6 +1048,157 @@ void PromiseTest::testAllAnyInitializerList()
 
 	QTRY_COMPARE(allPromise->state(), Deferred::Resolved);
 	QTRY_COMPARE(anyPromise->state(), Deferred::Resolved);
+}
+
+/*! \test Tests destruction of a Promise only.
+ */
+void PromiseTest::testPromiseDestruction()
+{
+	Deferred::Ptr deferred = Deferred::create();
+	QVariantList callbackCalls;
+	{
+		Promise::Ptr promise = Promise::create(deferred);
+		auto finalPromise = promise->always([&](const QVariant& data) {
+			callbackCalls.push_back(data);
+		});
+	}
+
+	QCOMPARE(callbackCalls.count(), 0);
+
+	deferred->resolve("foo");
+
+	/* Ensure that the callbacks are *NOT* called after the chain
+	 * has been destructed.
+	 */
+	QTest::qWait(100);
+	QCOMPARE(callbackCalls.count(), 0);
+}
+
+/*! \test Tests destruction of a complete Promise chain.
+ */
+void PromiseTest::testChainDestruction()
+{
+	Deferred::Ptr deferred = Deferred::create();
+	QVariantList callbackCalls;
+	{
+		Promise::Ptr finalPromise;
+		{
+			Promise::Ptr originalPromise = Promise::create(deferred);
+
+			finalPromise = originalPromise
+			->then(nullptr, nullptr, nullptr)
+			->then([&](const QVariant& data) {
+				callbackCalls.push_back(data);
+			}, [&](const QVariant& data) {
+				callbackCalls.push_back(data);
+			}, [&](const QVariant& data) {
+				callbackCalls.push_back(data);
+			});
+		}
+
+		// Allow potential callbacks to be called
+		QTest::qWait(100);
+
+		QVERIFY(callbackCalls.isEmpty());
+	}
+
+	// The chain should be *immediately* destroyed, so no qWait() here!
+	QVERIFY(callbackCalls.isEmpty());
+
+	deferred->resolve("foo");
+
+	/* Ensure that the callbacks are *NOT* called after the chain
+	 * has been destructed.
+	 */
+	QTest::qWait(100);
+	QVERIFY(callbackCalls.isEmpty());
+}
+
+/*! \test Tests destruction of a parent Deferred while a ChildDeferred
+ * is still holding a reference.
+ * This is testing behavior of the library in case of defective usage.
+ */
+void PromiseTest::testParentDeferredDestruction()
+{
+	/* We need to dynamically allocate the Deferred::Ptr and
+	 * explicitly leak it. This prevents that the QSharedPointer class
+	 * tries to delete the Deferred object after we have deleted it
+	 * because there still is a QSharedPointer holding a reference.
+	 * This way, we can delete the Deferred object without having the
+	 * test crash.
+	 */
+	Deferred::Ptr* deferred = new Deferred::Ptr(Deferred::create());
+
+	ChildDeferred::Ptr childDeferred = ChildDeferred::create(*deferred);
+
+	QCOMPARE(childDeferred->parents().size(), 1);
+
+	(*deferred)->resolve(); // Avoid warning
+
+	// This is buggy code which would lead to a crash in a real application
+	delete deferred->data(); // THIS IS AN ERROR! DON'T DO THIS IN A REAL APPLICATION!
+
+	QCOMPARE(childDeferred->parents().size(), 0);
+
+	childDeferred->resolve(); // Avoid warning
+}
+
+/*! Provides the data for the PromiseTest::testDelay() test.
+ */
+void PromiseTest::testDelay_data()
+{
+	QTest::addColumn<QString>("action");
+	QTest::addColumn<int>("delay");
+
+	//                                // action         // delay
+	QTest::newRow("deferred resolve") << ACTION_RESOLVE << 0;
+	QTest::newRow("deferred reject")  << ACTION_REJECT  << 0;
+	QTest::newRow("delayed resolve")  << ACTION_RESOLVE << 100;
+	QTest::newRow("delayed reject")   << ACTION_REJECT  << 100;
+}
+
+/*! \test Tests the Promise::delayedResolve() and Promise::delayedReject() methods.
+ */
+void PromiseTest::testDelay()
+{
+	QFETCH(QString, action);
+	QFETCH(int, delay);
+
+	Deferred::Ptr deferred = Deferred::create();
+	Promise::Ptr promise = Promise::create(deferred);
+
+	QVariant data = QVariant::fromValue(QString("foo bar"));
+	auto finalPromise = promise->then([=](const QVariant&) {
+		if (action == ACTION_RESOLVE)
+			return Promise::delayedResolve(data, delay);
+		else if (action == ACTION_REJECT)
+			return Promise::delayedReject(data, delay);
+		else
+			return Promise::createRejected("Unexpected action");
+	});
+
+	int delayDelta = delay * 0.3;
+
+	deferred->resolve("original data");
+
+	QCOMPARE(finalPromise->state(), Deferred::Pending);
+
+	QTest::qWait(delay - delayDelta + 1);
+
+	if (delay > 0)
+	{
+		QCOMPARE(finalPromise->state(), Deferred::Pending);
+
+		QTest::qWait(2 * delayDelta);
+	}
+
+	if (action == ACTION_RESOLVE)
+		QCOMPARE(finalPromise->state(), Deferred::Resolved);
+	else if (action == ACTION_REJECT)
+		QCOMPARE(finalPromise->state(), Deferred::Rejected);
+	else
+		QFAIL("Unexpected action");
+	QCOMPARE(finalPromise->data(), data);
 }